Create a generic read_ahead_driver
