+++ src/parallel/CMakeLists.txt
@@ -17,6 +17,7 @@ add_library_unity(
   pipeline_initialize_event.cpp
   pipeline_prepare_finish_event.cpp
   pipeline_schedule.cpp
+  scan_read_ahead.cpp
   task_executor.cpp
   task_notifier.cpp
   task_scheduler.cpp
