+++ src/parallel/scan_read_ahead.cpp
@@ -1,0 +1,42 @@
+#include "duckdb/parallel/scan_read_ahead.hpp"
+
+#include "duckdb/parallel/task_scheduler.hpp"
+
+namespace duckdb {
+
+void ReadAheadJobCompletion::FinishIOTask() {
+	const auto previous = pending_io_tasks.fetch_sub(1);
+	D_ASSERT(previous > 0);
+	if (previous > 1) {
+		// I/O tasks still outstanding, nothing to wake yet
+		return;
+	}
+	// wake the parked scan task, if any
+	const annotated_lock_guard<annotated_mutex> guard {parked_scan.lock};
+	parked_scan.UnblockTasks();
+}
+
+bool ReadAheadJobCompletion::TryPark(const InterruptState &interrupt_state) {
+	// checking the pending count under the same lock FinishIOTask takes before waking prevents lost wake-ups
+	const annotated_lock_guard<annotated_mutex> guard {parked_scan.lock};
+	if (pending_io_tasks.load() == 0) {
+		return false;
+	}
+	return parked_scan.BlockTask(interrupt_state);
+}
+
+void ReadAheadJobCompletion::WaitForIO() {
+	shared_ptr<Task> task;
+	while (pending_io_tasks.load() > 0) {
+		// run scheduled I/O inline, the tasks may belong to any job but always make progress
+		if (executor->GetTask(task)) {
+			task->Execute(TaskExecutionMode::PROCESS_ALL);
+			task.reset();
+			continue;
+		}
+		// the remaining I/O is running on other threads, wait for it to finish
+		TaskScheduler::YieldThread();
+	}
+}
+
+} // namespace duckdb
